# public/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "شركة المفتاح API",
    "description": "API للعقارات ونظام المستخدمين",
    "version": "1.0.0",
    "contact": {
      "name": "شركة المفتاح",
      "email": "info@almiftah.com"
    }
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8000/api",
      "description": "Local Server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
    "/auth/register": {
      "post": {
        "tags": ["Authentication"],
        "summary": "تسجيل مستخدم جديد",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password", "password_confirmation", "phone"],
                "properties": {
                  "name": {"type": "string", "example": "اسم المستخدم"},
                  "email": {"type": "string", "format": "email", "example": "user@example.com"},
                  "password": {"type": "string", "example": "password123"},
                  "password_confirmation": {"type": "string", "example": "password123"},
                  "phone": {"type": "string", "example": "966501234567"},
                  "role": {"type": "string", "enum": ["user", "owner"], "example": "user"}
                }
              }
            }
          }
        },
        "responses": {
          "201": {"description": "تم التسجيل بنجاح"},
          "422": {"description": "خطأ في التحقق"}
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}